+++ packages/block-library/src/query/edit/inspector-controls/taxonomy-controls.js
@@ -190,7 +190,6 @@ function TaxonomyItem( { taxonomy, termIds, onChange } ) {
 				displayTransform={ decodeEntities }
 				onChange={ onTermsChange }
 				__experimentalShowHowTo={ false }
-				__nextHasNoMarginBottom
 				__next40pxDefaultSize
 			/>
 		</div>