+++ packages/global-styles-ui/src/index.ts
@@ -5,3 +5,11 @@ export { TypographyVariations } from './typography-variations';
 
 // Ideally this should just be a core-data selector.
 export { default as useGlobalStylesRevisions } from './screen-revisions/use-global-styles-revisions';
+
+// Font Library exports
+export { FontLibrary } from './font-library/font-library';
+export { default as FontLibraryProvider } from './font-library/context';
+export { FontLibraryContext } from './font-library/context';
+export { default as InstalledFonts } from './font-library/installed-fonts';
+export { default as UploadFonts } from './font-library/upload-fonts';
+export { default as FontCollection } from './font-library/font-collection';